+++ test/fixtures/index.js
@@ -1,0 +1,3 @@
+module.exports = {
+	validParams: require('./validParams'),
+};